+++ examples/main.go
@@ -19,6 +19,8 @@ func main() {
 		main_iterate,
 		main_reduce,
 		main_sort,
+		main_seq_api,
+		main_seq2_api,
 	}
 
 	for _, f := range funcs {